backend/requirements.txt
flask
flask-cors
openai
httpx
python-dotenv
gunicorn